New feature: parsing keystrokes
The ActionElements containing keystrokes (Actions with type Keystroke
(actions that start with "press" or "presses")) are now parsed.

This means that typo's (for example "Downarrow" instead of "DownArrow") are
discovered by the compiler and will not cause error in Dragon. In addition
the user doesn't have to add spaces between keystrokes (spaces are optional).